+++ microprojects/calc/__init__.py
@@ -1,0 +1,2 @@
+from microprojects.calc.calculator import calc_main, calc
+from microprojects.calc.analyzer import lexical_analyzer, syntax_analyzer